pyproject.toml
[project]
name = "support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "fastapi[standard]>=0.121.3",
    "fastembed>=0.7.3",
    "httptools>=0.7.1",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.1.0",
    "langchain-postgres>=0.0.16",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "langgraph-checkpoint-postgres>=3.0.1",
    "pgvector>=0.2.5,<0.4.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pymupdf>=1.26.6",
    "python-jose[pycryptodome]>=3.5.0",
    "redis>=6.4.0",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.44",
    "supabase>=2.24.0",
    "uuid7>=0.1.0",
    "uvloop>=0.22.1",
]

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=0.1.1",
    "pytest-asyncio>=1.3.0",
]
